backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic-settings
python-jose[cryptography]
passlib[argon2]
httpx
passlib[argon2]
pydantic>=2.0
pydantic[email]  # <--- Esta linha foi adicionada/modificada
pytest
pytest-asyncio
bleach
google-generativeai>=0.3.0
mercadopago